Refactored handshake logic to its own class
DtlsSocket should be closer to role independent. The server-specific
handshake logic is now implemented in ServerHandshakeHandler. Later on
we can add ClientHandshakeHandler to take care of the client logic and
DtlsSocket can instantiate whichever of these it needs.

Cleaned the handshake logic a bit as well. It should be easier to follow
now and have less error conditions and stuff to track.

The one downside I'm a bit displeased about is handling the handshake
digest when sending messages. We can't call send_xxx() from the
handle_xxx() messages now as send_xxx() adds the handshakes to the
digest while the incoming digests are handled AFTER handle_xxx().

In future we might want to consider using setResponse to queue the
outgoing digests and then do the actual adding automatically in the
handle() method. However this would require moving the handshake
fragmentation to setResponse as well since we need to have the whole
unfragmented handshake buffers at hand when digesting them.