Client handshake done \o/
The library can now handshake in either client or server mode.

There's still a lot of stuff missing, mainly related to robustness. Plus
the conneciton crashes to DecryptFinal_ex:badDecrypt every now and then.
Could be due to renegotiate problems or something completely different.

But at least this is one milestone behind. \o/

Also implemented retransmission logic during the client handshake. It's
not perfect and needs refactoring, but at least we got working logic
somewhere now.

The retransmission works by the client saving the last packets it sent
and setting up a timer. If there's no state change the timer will
trigger a retransmission.

Unfortunately this approach needs some hacking in DtlsRecordLayer as
well. Ideally we'd have the retransmission triggered by the handshake
handlers but handled completely in the RecordLayer. This probably needs
for the RecordLayer to store the flight and return a flight ID which the
handshake layer can use to trigger the retransmit.

Also we want to refactor the common parts of the handshake handlers into
one class (setResponse, handleHandshake, etc). Currently there's quite a
bit duplicated code between these two.