Combining messages into one UDP frame if possible
DtlsRecordLayer tries to combine multiple messages into one UDP frame as
long as there is space in it. This especially speeds up handshake in
cases where clients ignore too early packets as it prevents the later
packets from arriving early.

(Fine. Real reason was the bugs in our own implementation which broke
the sockets if packets from the next epoch arrived before the epoch was
finished)