SequenceNumber returns reference to counter
Instead of creating a copy of the sequence counter, we return a
reference from the next() method in SequenceNumber. The packet to which
the counter value is assigned is turned into a buffer soon afterwards so
there's no second call of next() which makes this usage safe.

The benefit of the reference is that we don't always need to
re-construct the sequence number buffer when returning it from the
next():